+++ src/Observers/EntityAttributeUpdateObserver.php
@@ -1,0 +1,76 @@
+<?php
+
+/**
+ * TechDivision\Import\Attribute\Observers\EntityAttributeUpdateObserver
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Attribute\Observers;
+
+use TechDivision\Import\Attribute\Utils\MemberNames;
+
+/**
+ * Observer that update's the EAV entity attribute itself.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+class EntityAttributeUpdateObserver extends EntityAttributeObserver
+{
+
+    /**
+     * Initialize the attribute with the passed attributes and returns an instance.
+     *
+     * @param array $attr The attribute attributes
+     *
+     * @return array The initialized attribute
+     */
+    protected function initializeAttribute(array $attr)
+    {
+
+        // load the attributes
+        $entityTypeId = $attr[MemberNames::ENTITY_TYPE_ID];
+        $attributeId = $attr[MemberNames::ATTRIBUTE_ID];
+        $attributeSetId = $attr[MemberNames::ATTRIBUTE_SET_ID];
+        $attributeGroupId = $attr[MemberNames::ATTRIBUTE_GROUP_ID];
+
+        // try to load the entity attribute with the IDs
+        if ($entityAttribute = $this->loadEntityAttributeByEntityTypeAndAttributeIdAndAttributeSetIdAndAttributeGroupId($entityTypeId, $attributeId, $attributeSetId, $attributeGroupId)) {
+            return $this->mergeEntity($entityAttribute, $attr);
+        }
+
+        // simply return the attributes
+        return $attr;
+    }
+
+    /**
+     * Return's the EAV entity attribute with the passed entity type, attribute, attribute set and attribute group ID.
+     *
+     * @param integer $entityTypeId     The ID of the EAV entity attribute's entity type to return
+     * @param integer $attributeId      The ID of the EAV entity attribute's attribute to return
+     * @param integer $attributeSetId   The ID of the EAV entity attribute's attribute set to return
+     * @param integer $attributeGroupId The ID of the EAV entity attribute's attribute group to return
+     *
+     * @return array The EAV entity attribute
+     */
+    protected function loadEntityAttributeByEntityTypeAndAttributeIdAndAttributeSetIdAndAttributeGroupId($entityTypeId, $attributeId, $attributeSetId, $attributeGroupId)
+    {
+        return $this->getAttributeBunchProcessor()->loadEntityAttributeByEntityTypeAndAttributeIdAndAttributeSetIdAndAttributeGroupId($entityTypeId, $attributeId, $attributeSetId, $attributeGroupId);
+    }
+}