+++ src/Utils/MemberNames.php
@@ -290,4 +290,11 @@ class MemberNames extends \TechDivision\Import\Utils\MemberNames
      * @var string
      */
     const TYPE = 'type';
+
+    /**
+     * Name for the member 'attribute_group_id'.
+     *
+     * @var string
+     */
+    const ATTRIBUTE_GROUP_ID = 'attribute_group_id';
 }