+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Update generic configuration file
 * Add functionality to persist attribute labels
+* Add functionality to persist entity attribute data
 * Implement EAV attribute option swatch delete/add-update + replace functionality
 * Add handling for additional_data attribute => necessary to create swatch attributes